+++ openhab_rules/warmup.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+puts 'Starting Warmup'
+
 require 'openhab'
 
 logger.info('OpenHAB warmup complete')