+++ CHANGELOG.md
@@ -1,10 +1,38 @@
-# JRuby OpenHAB Scripting change log
+JRuby OpenHAB Scripting Change Log
 
-All notable changes to this project will be documented in this file.
+# [2.12.0](https://github.com/boc-tothefuture/openhab-jruby/compare/2.11.1...2.12.0) (2021-02-02)
 
-Project versioning adheres to [Semantic Versioning](http://semver.org/). <br>
-Commit convention is based on [Conventional Commits](http://conventionalcommits.org). <br>
-Change log format is based on [Keep a Changelog](http://keepachangelog.com/). <br>
+
+### Bug Fixes
+
+* return nil for items['nonexistent'] instead of raising an exception ([4a412f8](https://github.com/boc-tothefuture/openhab-jruby/commit/4a412f81daf46c86469d8badeac2327a6c1816d3))
+
+
+### Features
+
+* add Item.include? to check for item's existence ([1a8fd3a](https://github.com/boc-tothefuture/openhab-jruby/commit/1a8fd3aad11fb0549dc2c7308f26946ffb8e899c))
+
+## [2.11.1](https://github.com/boc-tothefuture/openhab-jruby/compare/2.11.0...2.11.1) (2021-02-01)
+
+
+### Bug Fixes
+
+* **group:**  support for accessing triggering item in group updates ([6204f0a](https://github.com/boc-tothefuture/openhab-jruby/commit/6204f0a8f33e08abddcc130b46a2fe39c5f4bb49))
+
+# [2.11.0](https://github.com/boc-tothefuture/openhab-jruby/compare/2.10.1...2.11.0) (2021-02-01)
+
+
+### Features
+
+* Add Duration.to_s ([b5b9c81](https://github.com/boc-tothefuture/openhab-jruby/commit/b5b9c8176f995ad996ac481c4c23c614bd5f54f7))
+
+## 2.10.0
+### Changed
+- Library now released as a Ruby Gem
+
+## 2.9.0
+### Added
+- Support OpenHAB Actions
 
 ## 2.8.1
 ### Fixed